% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_add_input.R
\name{prm_add_input}
\alias{prm_add_input}
\title{Add input to model input data frame}
\usage{
prm_add_input(.input_df, ...)
}
\arguments{
\item{.input_df}{a data frame with model input information as created by
\link{prm_create_input_df}}

\item{...}{parameters passed to \link{prm_create_input_df}}
}
\description{
Add input to model input data frame
}
