% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_add_out_df.R
\name{prm_add_out_df}
\alias{prm_add_out_df}
\title{Add output data frame}
\usage{
prm_add_out_df(.expmt, prioritize_files = c("Summary.OUT", "PlantGro.OUT"))
}
\arguments{
\item{.expmt}{a data frame with information about experiments to be used for
estimation}

\item{prioritize_files}{a character vector of files to prioritize in
searching for output variables}
}
\description{
Add output data frame
}
