% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_add_prm.R
\name{prm_add_prm}
\alias{prm_add_prm}
\title{Add a parameter to a parameter estimation data frame}
\usage{
prm_add_prm(.prm_df, ...)
}
\arguments{
\item{.prm_df}{a data frame that contains parameter information as generated
by \link{prm_create_prm_df}}

\item{...}{arguments passed to \link{prm_create_prm_df} (see
\link{prm_create_prm_df} for details)}
}
\description{
Add a parameter to a parameter estimation data frame
}
