% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_add_transform.R
\name{prm_add_transform}
\alias{prm_add_transform}
\title{Create a parameter transformation}
\usage{
prm_add_transform(prm_df, ptrans, pfile = "", ...)
}
\description{
Create a parameter transformation
}
