% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_add_variances.R
\name{prm_add_variances}
\alias{prm_add_variances}
\title{Add hyperparameters for variance terms ("model error") to prm_df}
\usage{
prm_add_variances(
  prm_df,
  pmin = 0,
  pmax = Inf,
  pmu = 0,
  psigma = 1,
  pdist = "normal",
  ...
)
}
\arguments{
\item{prm_df}{a data frame as created by \link{prm_create_prm_df}}

\item{pmin}{a numeric vector providing the lower bound(s) for parameter(s) to
be estimated. The default is -Inf (i.e. unbounded).}

\item{pmax}{a numeric vector providing the upper bound(s) for parameter(s) to
be estimated. The default is Inf (i.e. unbounded)}

\item{pmu}{a numeric vector providing the mean(s) of the prior
distribution(s) for parameter(s) to be estimated}

\item{psigma}{a numeric vector providing the square root(s) of the
variance(s) of the prior distribution(s) for parameter(s) to be estimated}

\item{pdist}{a character vector providing the prior distribution(s) for
parameter(s) to be estimated (one of "uniform" of "normal")}
}
\description{
Add hyperparameters for variance terms ("model error") to prm_df
}
