% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_calc_pval.R
\name{prm_calc_pval}
\alias{prm_calc_pval}
\title{Calculate parameter values (regular and transformed)}
\usage{
prm_calc_pval(prm_df, pval_in = NULL, n = 1)
}
\description{
Calculate parameter values (regular and transformed)
}
