% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_expmt.R
\name{prm_create_expmt}
\alias{prm_create_expmt}
\title{Create an experiment for parameter estimation}
\usage{
prm_create_expmt(expmt, data)
}
\arguments{
\item{expmt}{either the file name of a DSSAT-formatted experiment
file (File X) or a list as would be returned by
\link[DSSAT]{read_filex} or \link[DSSAT]{filex_template}.}

\item{data}{a data frame of seasonal and within-season data to use for
parameter estimation or a character vector of names for DSSAT-formatted
files for seasonal (File A) or within-season (File T) data}
}
\description{
Create an experiment for parameter estimation
}
