% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_expmt_df.R
\name{prm_create_expmt_df}
\alias{prm_create_expmt_df}
\title{Create a table of experiments for parameter estimation}
\usage{
prm_create_expmt_df(expmt, trno = NULL, variables = NULL)
}
\arguments{
\item{expmt}{either a vector of file names of DSSAT-formatted experiment
files (File X) or a list each element of which is a list as would be
returned by \link[DSSAT]{read_filex} or \link[DSSAT]{filex_template}.}

\item{trno}{an optional list of numeric vectors of treatment numbers to be
used for the parameter estimation}

\item{variables}{an optional list of character vectors of variable names to be used
for the parameter estimation}
}
\description{
Create a table of experiments for parameter estimation
}
