% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_expmt_tbl.R
\name{prm_create_expmt_tbl}
\alias{prm_create_expmt_tbl}
\title{Create a table of experiments for parameter estimation}
\usage{
prm_create_expmt_tbl(
  expmt,
  trno = NULL,
  data_types = NULL,
  rewrite_filex = FALSE
)
}
\arguments{
\item{expmt}{either a vector of file names of DSSAT-formatted experiment
files (File X) or a list each element of is a list as would be returned by
\link[DSSAT]{read_filex} or \link[DSSAT]{filex_template}.}

\item{data_types}{a list of character vectors of variable names to be used
for the parameter estimation}

\item{rewrite_filex}{a logical value indicating whether to rewrite the
experiment}
}
\description{
Create a table of experiments for parameter estimation
}
