% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_obj_fun.R
\name{prm_create_obj_fun}
\alias{prm_create_obj_fun}
\title{Create an objective function}
\usage{
prm_create_obj_fun(obj, aggr, by)
}
\arguments{
\item{obj}{an expression or formula calculating the value of the objective
function in which the observed data are named \code{obs} and the
corresponding simulated output data are named \code{yhat}}

\item{aggr}{an expression or formula object for aggregating objective
function values into a single scalar result in which objective function
values are named \code{obj_val}}

\item{by}{an optional formula providing a set of grouping factors by which to
group the initial calculation of the objective function value. For more
details, see the \code{f} argument of the \link{split} method for data
frames.}
}
\description{
Create an objective function
}
