% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_obs_df.R
\name{prm_create_obs_df}
\alias{prm_create_obs_df}
\title{Create a table of observations for parameter estimation}
\usage{
prm_create_obs_df(data_df, time_alias, group_cols)
}
\arguments{
\item{data_df}{a data frame of seasonal summary and/or within-season data to
a column for time (named \code{time} or another name as specified in the
\code{time_alias} argument). Columns with grouping variables should be
listed in the \code{group_cols} argument. The grouping variables should
match those provided to \link{prm_create_sim} or \link{prm_create_sim_df}.}

\item{time_alias}{an optional expression, providing a new name to use
for the time variable. This name should match what the model returns.}

\item{group_cols}{an expression, indicating columns in \code{data_df}
containing grouping variables.}
}
\value{
a data frame in long format that includes columns for time
(named \code{time} or another name as specified in the \code{time_alias}
argument), variable name (named \code{variable}) and observed data (named
\code{obs}).
}
\description{
Create a table of observations for parameter estimation
}
