% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_prm_df.R
\name{prm_create_prm_df}
\alias{prm_create_prm_df}
\title{Create a parameter data frame for estimation}
\usage{
prm_create_prm_df(
  pname,
  pfile = NA,
  pmin = -Inf,
  pmax = Inf,
  pmu = NA_real_,
  psigma = NA_real_,
  pdist = "uniform",
  ptier = NA_character_,
  pkey = NA_character_,
  plev = NA_integer_,
  pind = NA_integer_,
  pfmt = NA_character_,
  pnum = NULL,
  pwt = NULL
)
}
\arguments{
\item{pname}{a character vector providing the name(s) of parameter(s) to be
estimated}

\item{pfile}{a character vector providing the file name(s) of the
parameter(s) to be estimated}

\item{pmin}{a numeric vector providing the lower bound(s) for parameter(s) to
be estimated. The default is -Inf (i.e. unbounded).}

\item{pmax}{a numeric vector providing the upper bound(s) for parameter(s) to
be estimated. The default is Inf (i.e. unbounded)}

\item{pmu}{a numeric vector providing the mean(s) of the prior
distribution(s) for parameter(s) to be estimated}

\item{psigma}{a numeric vector providing the square root(s) of the
variance(s) of the prior distribution(s) for parameter(s) to be estimated}

\item{pdist}{a character vector providing the prior distribution(s) for
parameter(s) to be estimated (one of "uniform" of "normal")}

\item{ptier}{an optional character vector providing the}

\item{pkey}{an optional character vector providing the key value to use for
matching to an entry within \code{pfile}}
}
\description{
Create a parameter data frame for estimation
}
