% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_prm_est.R
\name{prm_create_prm_est}
\alias{prm_create_prm_est}
\title{Create a parameter estimation object}
\usage{
prm_create_prm_est(sim_df, prm_df, obs_df, run_fun, smry_fun, obj_fun)
}
\arguments{
\item{sim_df}{a data frame of simulation definitions as created by
\link{prm_create_sim_df}}

\item{prm_df}{a data frame of parameter definitions as created by
\link{prm_create_prm_df}}

\item{obs_df}{a data frame of observed data as created by
\link{prm_create_obs_df}}

\item{run_fun}{a function that takes a parameter vector as its first
argument and a prm_est object as its second argument that calls the model
and returns model output in the form of a data frame corresponding to
\code{obs_df} with model simulated values in column \code{yhat}}

\item{obj_fun}{a function defining the objective function that should be
minimized during the parameter estimation as defined by
\link{prm_create_obj_fun}. This function should accept as arguments at
minimum two data frames: one for observations (\code{obs_df} and one for
output of simulations (\code{yhat_df}). The function should return a scalar
value.

@param smry_fun an optional function as created by
\link{prm_create_smry_fun} that computes summary variables based on direct
model outputs}
}
\description{
Create a parameter estimation object
}
