% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_prm_sa.R
\name{prm_create_prm_sa}
\alias{prm_create_prm_sa}
\title{Create a parameter estimation object}
\usage{
prm_create_prm_sa(
  expmt_df,
  inp_df,
  prm_df,
  sa_fun,
  model_type = "DSSAT-CSM",
  model_call
)
}
\arguments{
\item{expmt_df}{a data frame of experiment definitions as created by
\link{prm_create_expmt_df}}

\item{prm_df}{a data frame}

\item{input_df}{a data frame of input definitions as created by
\link{prm_create_input_df}}
}
\description{
Create a parameter estimation object
}
