% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_run_fun.R
\name{prm_create_run_fun}
\alias{prm_create_run_fun}
\title{Create a function for running a model}
\usage{
prm_create_run_fun(fun_name, prm_name = "prm", time_name = "time")
}
\arguments{
\item{fun_name}{a character string providing the name of the function to be
called for running each simulation}

\item{prm_name}{a character string providing the name of the parameter
vector}

\item{time_name}{a character string providing the name of the time variable
in the model output}
}
\description{
Create a function for running a model
}
