% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_sim_df.R
\name{prm_create_sim_df}
\alias{prm_create_sim_df}
\title{Create a table of simulations for parameter estimation}
\usage{
prm_create_sim_df(sim_args, ...)
}
\arguments{
\item{sim_args}{a list of arguments to be used when running simulations with
the model function. Each argument should be supplied as a list. To use the
same argument value for all simulations, the list should be of length one.
Otherwise, the length of the list should correspond to the number of
simulations}

\item{...}{optional additional arguments providing grouping factors for
identifying the simulation}
}
\description{
Create a table of simulations for parameter estimation
}
