% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_smry_fun.R
\name{prm_create_smry_fun}
\alias{prm_create_smry_fun}
\title{Create a function for summarizing model outputs}
\usage{
prm_create_smry_fun(prm_name = "prm", time_name = "time", ...)
}
\arguments{
\item{prm_name}{a character string providing the name of the parameter
vector in the exressions}

\item{time_name}{a character string providing the name of the time variable
in the model output}

\item{...}{named expressions each of which can be evaluated with model
outputs to generate summary output}
}
\description{
Create a function for summarizing model outputs
}
