% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_est_DEoptim.R
\name{prm_est_DEoptim}
\alias{prm_est_DEoptim}
\title{Use the DEoptim package to perform global parameter optimization}
\usage{
prm_est_DEoptim(prm_est, control = DEoptim::DEoptim.control())
}
\arguments{
\item{prm_est}{a named list created using \link{prm_create_prm_est}}

\item{control}{a named list created using \link[DEoptim]{DEoptim.control}() from the
DEoptim package (See \link[DEoptim]{DEoptim.control} for details)}
}
\description{
Use the DEoptim package to perform global parameter optimization
}
