% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_est_optim.R
\name{prm_est_optim}
\alias{prm_est_optim}
\title{Use the DEoptim package to perform global parameter optimization}
\usage{
prm_est_optim(prm_est, method = "L-BFGS-B", control = list())
}
\arguments{
\item{prm_est}{a named list created using \link{prm_create_prm_est}}

\item{control}{a named list with options for \link[stats]{optim}() from the
\code{stats} package (See \link[stats]{optim} for details)}
}
\description{
Use the DEoptim package to perform global parameter optimization
}
