% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_get_model_outputs.R
\name{prm_get_model_outputs}
\alias{prm_get_model_outputs}
\title{Run model and return model outputs}
\usage{
prm_get_model_outputs(prm_est, pvals)
}
\arguments{
\item{prm_est}{a list object as returned by prm_create_prm_est()}

\item{pvals}{a vector of parameter values for which to generate model outputs}
}
\description{
Run model and return model outputs
}
