% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_sa_morris.R
\name{prm_sa_morris}
\alias{prm_sa_morris}
\title{Use the Morris method to perform global sensitivity analysis}
\usage{
prm_sa_morris(prm_sa, ...)
}
\arguments{
\item{prm_sa}{a named list created using \link{prm_create_prm_sa}}

\item{...}{additional arguments passed on to the \code{morris} function.
See \link[sensitivity]{morris}() for more details.}
}
\description{
Use the Morris method to perform global sensitivity analysis
}
