% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DEoptim_estimation.R
\name{run_DEoptim_estimation}
\alias{run_DEoptim_estimation}
\title{Use the DEoptim package to perform global parameter optimization}
\usage{
run_DEoptim_estimation(prm_est, control = DEoptim::DEoptim.control())
}
\arguments{
\item{prm_est}{a named list created using \link{create_prm_est}}

\item{control}{a named list created using \link[DEoptim]{DEoptim.control}() from the
DEoptim package (See \link[DEoptim]{DEoptim.control} for details)}
}
\description{
Use the DEoptim package to perform global parameter optimization
}
