% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_add_variances.R
\name{prm_add_variances}
\alias{prm_add_variances}
\title{Add hyperparameters for variance terms ("model error") to prm_tbl}
\usage{
prm_add_variances(
  prm_tbl,
  obs_tbl,
  pmin = 0,
  pmax = Inf,
  pmu = 0,
  psigma = 1,
  pdist = "normal",
  ...
)
}
\arguments{
\item{obs_tbl}{a tibble as produced by the create_dssat_expmt()
function grouped by the factors which identify the observation
groups for which a variance term should be estimated}
}
\description{
Add hyperparameters for variance terms ("model error") to prm_tbl
}
