% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_prm_df.R
\name{prm_create_prm_df}
\alias{prm_create_prm_df}
\title{Create a parameter data frame for estimation}
\usage{
prm_create_prm_df(
  pname,
  pmin = -Inf,
  pmax = Inf,
  pmu = NA_real_,
  psigma = NA_real_,
  pdist = "uniform",
  pfmt = NA_character_,
  ...
)
}
\arguments{
\item{pname}{a character vector providing the name(s) of parameter(s) to be
estimated}

\item{pmin}{a numeric vector providing the lower bound(s) for parameter(s) to
be estimated. The default is -Inf (i.e. unbounded).}

\item{pmax}{a numeric vector providing the upper bound(s) for parameter(s) to
be estimated. The default is Inf (i.e. unbounded)}

\item{pmu}{a numeric vector providing the mean(s) of the prior
distribution(s) for parameter(s) to be estimated}

\item{psigma}{a numeric vector providing the square root(s) of the
variance(s) of the prior distribution(s) for parameter(s) to be estimated}

\item{pdist}{a character vector providing the prior distribution(s) for
parameter(s) to be estimated (one of "uniform" of "normal")}

\item{...}{other arguments used to uniquely identify the parameter}
}
\description{
Create a parameter data frame for estimation
}
