% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_prm_est.R
\name{prm_create_prm_est}
\alias{prm_create_prm_est}
\title{Create a parameter estimation object}
\usage{
prm_create_prm_est(
  sim_df,
  prm_df,
  obj_fun,
  call_model,
  model_type = "DSSAT-CSM",
  input_df
)
}
\arguments{
\item{sim_df}{a data frame of simulation definitions as created by
\link{prm_create_sim_df}}

\item{prm_df}{a data frame of parameter definitions as created by
\link{prm_create_prm_df}}

\item{obj_fun}{a function defining the objective function that should be
minimized during the parameter estimation as defined by
\link{prm_create_obj_fun}. This function should accept as arguments at
minimum two data frames: one for observations (\code{obs_df} and one for
output of simulations (\code{yhat_df}). The function should return a scalar
value.

@param call_model a function that calls}

\item{input_df}{an optional data frame of input definitions as created by
\link{prm_create_input_df} for models with external input files}
}
\description{
Create a parameter estimation object
}
