% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_prm_est.R
\name{prm_create_prm_est}
\alias{prm_create_prm_est}
\title{Create a parameter estimation object}
\usage{
prm_create_prm_est(
  expmt_df,
  inp_df,
  prm_df,
  obj_fun,
  model_type = "DSSAT-CSM",
  dssat_call
)
}
\arguments{
\item{expmt_df}{a data frame of experiment definitions as created by
\link[csmparameter]{prm_create_expmt_df}}

\item{inp_df}{a data frame of input definitions as created by
\link{prm_create_inp_df}}

\item{prm_df}{a data frame}
}
\description{
Create a parameter estimation object
}
