% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_prm_est.R
\name{prm_create_prm_est}
\alias{prm_create_prm_est}
\title{Create a parameter estimation object}
\usage{
prm_create_prm_est(
  expmt_tbl,
  input_tbl,
  prm_tbl,
  obj_fun,
  model_type = "DSSAT-CSM",
  dssat_call
)
}
\arguments{
\item{expmt_tbl}{a data frame of experiment definitions as created by
\link[csmparameter]{prm_create_expmt_tbl}}

\item{input_tbl}{a data frame of inputs}
}
\description{
Create a parameter estimation object
}
