% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_create_sim.R
\name{prm_create_sim}
\alias{prm_create_sim}
\title{Create a simulation for parameter estimation}
\usage{
prm_create_sim(sim_args, ...)
}
\arguments{
\item{sim_args}{a list of arguments to be used when running the simulation
with the model function}

\item{...}{optional additional arguments providing grouping factors for
identifying the simulation}
}
\description{
Create a simulation for parameter estimation
}
